Split fontframe.py into multiple files for #8
